pyproject.toml
[tool.poetry]
name = "orientseq"
version = "0.2.1"
description = "Orients fa/fq/sam/bam file: reverse revcomp by finding polyA tails"
authors = ["Dmitry Solovey <dsolovey@ist.ac.at>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
biopython = "^1.84"
pysam = "^0.22.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
orient = 'orientseq.__main__:main'
